WIP: Integrate drag-and-drop in BlockLibrary
