+++ src/app/MainContent.tsx
@@ -1,16 +1,19 @@
 import { Stack } from "@mui/material";
 import { TitledSection } from "./TitledSection.tsx";
 import { BlockLibrary } from "../block-library/BlockLibrary.tsx";
-import { SolutionBox } from "../solution-box/SolutionBox.tsx";
+import {
+  SolutionBox,
+  type SolutionBoxProps,
+} from "../solution-box/SolutionBox.tsx";
 
-export function MainContent() {
+export function MainContent(solutionBoxProps: SolutionBoxProps) {
   return (
     <Stack direction={"row"} spacing={2}>
       <TitledSection title={"Block Library"}>
         <BlockLibrary />
       </TitledSection>
       <TitledSection title={"Solution Box"}>
-        <SolutionBox />
+        <SolutionBox {...solutionBoxProps} />
       </TitledSection>
     </Stack>
   );
